Network.AWS.MarketplaceMetering.MeterUsage

Description

API to emit metering records. For identical requests, the API is idempotent. It simply returns the metering record ID.

MeterUsage is authenticated on the buyer's AWS account, generally when running from an EC2 instance on the AWS Marketplace.

Creating a Request

meterUsage

Arguments

:: Text
muProductCode

-> UTCTime
muTimestamp

-> Text
muUsageDimension

-> Natural
muUsageQuantity

-> Bool
muDryRun

-> MeterUsage

Creates a value of MeterUsage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

muProductCode - Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

muTimestamp - Timestamp of the hour, recorded in UTC. The seconds and milliseconds portions of the timestamp will be ignored.

muUsageDimension - It will be one of the fcp dimension name provided during the publishing of the product.

muUsageQuantity - Consumption value for the hour.

muDryRun - Checks whether you have the permissions required for the action, but does not make the request. If you have the permissions, the request returns DryRunOperation; otherwise, it returns UnauthorizedException.

data MeterUsage

See: meterUsage smart constructor.

Request Lenses

muProductCode :: Lens' MeterUsage Text

Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.

muTimestamp :: Lens' MeterUsage UTCTime

Timestamp of the hour, recorded in UTC. The seconds and milliseconds portions of the timestamp will be ignored.

muUsageDimension :: Lens' MeterUsage Text

It will be one of the fcp dimension name provided during the publishing of the product.

muUsageQuantity :: Lens' MeterUsage Natural

Consumption value for the hour.

muDryRun :: Lens' MeterUsage Bool

Checks whether you have the permissions required for the action, but does not make the request. If you have the permissions, the request returns DryRunOperation; otherwise, it returns UnauthorizedException.

Destructuring the Response

meterUsageResponse

Arguments

:: Int
mursResponseStatus

-> MeterUsageResponse

Creates a value of MeterUsageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mursMeteringRecordId - Undocumented member.

mursResponseStatus - -- | The response status code.

data MeterUsageResponse

See: meterUsageResponse smart constructor.

Response Lenses

mursMeteringRecordId :: Lens' MeterUsageResponse (Maybe Text)

Undocumented member.

mursResponseStatus :: Lens' MeterUsageResponse Int

- | The response status code.