Network.AWS.MarketplaceMetering.ResolveCustomer

Description

ResolveCustomer is called by a SaaS application during the registration process. When a buyer visits your website during the registration process, the buyer submits a registration token through their browser. The registration token is resolved through this API to obtain a CustomerIdentifier and product code.

Creating a Request

resolveCustomer

Arguments

:: Text
rcRegistrationToken

-> ResolveCustomer

Creates a value of ResolveCustomer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcRegistrationToken - When a buyer visits your website during the registration process, the buyer submits a registration token through the browser. The registration token is resolved to obtain a CustomerIdentifier and product code.

data ResolveCustomer

Contains input to the ResolveCustomer operation.

See: resolveCustomer smart constructor.

Request Lenses

rcRegistrationToken :: Lens' ResolveCustomer Text

When a buyer visits your website during the registration process, the buyer submits a registration token through the browser. The registration token is resolved to obtain a CustomerIdentifier and product code.

Destructuring the Response

resolveCustomerResponse

Arguments

:: Int
rcrsResponseStatus

-> ResolveCustomerResponse

Creates a value of ResolveCustomerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rcrsCustomerIdentifier - The CustomerIdentifier is used to identify an individual customer in your application. Calls to BatchMeterUsage require CustomerIdentifiers for each UsageRecord.

rcrsProductCode - The product code is returned to confirm that the buyer is registering for your product. Subsequent BatchMeterUsage calls should be made using this product code.

rcrsResponseStatus - -- | The response status code.

data ResolveCustomerResponse

The result of the ResolveCustomer operation. Contains the CustomerIdentifier and product code.

See: resolveCustomerResponse smart constructor.

Response Lenses

rcrsCustomerIdentifier :: Lens' ResolveCustomerResponse (Maybe Text)

The CustomerIdentifier is used to identify an individual customer in your application. Calls to BatchMeterUsage require CustomerIdentifiers for each UsageRecord.

rcrsProductCode :: Lens' ResolveCustomerResponse (Maybe Text)

The product code is returned to confirm that the buyer is registering for your product. Subsequent BatchMeterUsage calls should be made using this product code.

rcrsResponseStatus :: Lens' ResolveCustomerResponse Int

- | The response status code.