Network.AWS.MarketplaceMetering.Types

Description

Service Configuration

marketplaceMetering :: Service

API version 2016-01-14 of the Amazon Marketplace Metering SDK configuration.

Errors

_InvalidEndpointRegionException :: AsError a => Getting (First ServiceError) a ServiceError

The endpoint being called is in a region different from your EC2 instance. The region of the Metering service endpoint and the region of the EC2 instance must match.

_InvalidProductCodeException :: AsError a => Getting (First ServiceError) a ServiceError

The product code passed does not match the product code used for publishing the product.

_InvalidUsageDimensionException :: AsError a => Getting (First ServiceError) a ServiceError

The usage dimension does not match one of the UsageDimensions associated with products.

_DuplicateRequestException :: AsError a => Getting (First ServiceError) a ServiceError

A metering record has already been emitted by the same EC2 instance for the given {usageDimension, timestamp} with a different usageQuantity.

_TimestampOutOfBoundsException :: AsError a => Getting (First ServiceError) a ServiceError

The timestamp value passed in the meterUsage() is out of allowed range.

_ThrottlingException :: AsError a => Getting (First ServiceError) a ServiceError

The calls to the MeterUsage API are throttled.

_InternalServiceErrorException :: AsError a => Getting (First ServiceError) a ServiceError

An internal error has occurred. Retry your request. If the problem persists, post a message with details on the AWS forums.

_InvalidTokenException :: AsError a => Getting (First ServiceError) a ServiceError

Prism for InvalidTokenException' errors.

_ExpiredTokenException :: AsError a => Getting (First ServiceError) a ServiceError

The submitted registration token has expired. This can happen if the buyer's browser takes too long to redirect to your page, the buyer has resubmitted the registration token, or your application has held on to the registration token for too long. Your SaaS registration website should redeem this token as soon as it is submitted by the buyer's browser.

_InvalidCustomerIdentifierException :: AsError a => Getting (First ServiceError) a ServiceError

You have metered usage for a CustomerIdentifier that does not exist.

UsageRecordResultStatus

data UsageRecordResultStatus

Constructors

CustomerNotSubscribed

DuplicateRecord

Success

UsageRecord

data UsageRecord

A UsageRecord indicates a quantity of usage for a given product, customer, dimension and time.

Multiple requests with the same UsageRecords as input will be deduplicated to prevent double charges.

See: usageRecord smart constructor.

usageRecord

Arguments

:: UTCTime
urTimestamp

-> Text
urCustomerIdentifier

-> Text
urDimension

-> Natural
urQuantity

-> UsageRecord

Creates a value of UsageRecord with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urTimestamp - Timestamp of the hour, recorded in UTC. The seconds and milliseconds portions of the timestamp will be ignored. Your application can meter usage for up to one hour in the past.

urCustomerIdentifier - The CustomerIdentifier is obtained through the ResolveCustomer operation and represents an individual buyer in your application.

urDimension - During the process of registering a product on AWS Marketplace, up to eight dimensions are specified. These represent different units of value in your application.

urQuantity - The quantity of usage consumed by the customer for the given dimension and time.

urTimestamp :: Lens' UsageRecord UTCTime

Timestamp of the hour, recorded in UTC. The seconds and milliseconds portions of the timestamp will be ignored. Your application can meter usage for up to one hour in the past.

urCustomerIdentifier :: Lens' UsageRecord Text

The CustomerIdentifier is obtained through the ResolveCustomer operation and represents an individual buyer in your application.

urDimension :: Lens' UsageRecord Text

During the process of registering a product on AWS Marketplace, up to eight dimensions are specified. These represent different units of value in your application.

urQuantity :: Lens' UsageRecord Natural

The quantity of usage consumed by the customer for the given dimension and time.

UsageRecordResult

data UsageRecordResult

A UsageRecordResult indicates the status of a given UsageRecord processed by BatchMeterUsage.

See: usageRecordResult smart constructor.

usageRecordResult :: UsageRecordResult

Creates a value of UsageRecordResult with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

urrStatus - The UsageRecordResult Status indicates the status of an individual UsageRecord processed by BatchMeterUsage. * Success - The UsageRecord was accepted and honored by BatchMeterUsage. * CustomerNotSubscribed - The CustomerIdentifier specified is not subscribed to your product. The UsageRecord was not honored. Future UsageRecords for this customer will fail until the customer subscribes to your product. * DuplicateRecord - Indicates that the UsageRecord was invalid and not honored. A previously metered UsageRecord had the same customer, dimension, and time, but a different quantity.

urrUsageRecord - The UsageRecord that was part of the BatchMeterUsage request.

urrMeteringRecordId - The MeteringRecordId is a unique identifier for this metering event.

urrStatus :: Lens' UsageRecordResult (Maybe UsageRecordResultStatus)

The UsageRecordResult Status indicates the status of an individual UsageRecord processed by BatchMeterUsage. * Success - The UsageRecord was accepted and honored by BatchMeterUsage. * CustomerNotSubscribed - The CustomerIdentifier specified is not subscribed to your product. The UsageRecord was not honored. Future UsageRecords for this customer will fail until the customer subscribes to your product. * DuplicateRecord - Indicates that the UsageRecord was invalid and not honored. A previously metered UsageRecord had the same customer, dimension, and time, but a different quantity.

urrUsageRecord :: Lens' UsageRecordResult (Maybe UsageRecord)

The UsageRecord that was part of the BatchMeterUsage request.

urrMeteringRecordId :: Lens' UsageRecordResult (Maybe Text)

The MeteringRecordId is a unique identifier for this metering event.